requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.25.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.12.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: for local LLM support
# ollama>=0.1.0
# llama-cpp-python>=0.2.0

# System monitoring
psutil>=5.9.0